one more refactoring - a little callback parameter to a callback function. piece of cake :)
